Pen Test

+++ src/routes/franchiseRouter.js
@@ -119,6 +119,9 @@ franchiseRouter.delete(
   authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
+    if (!user.isRole(Role.Admin)) {
+      return res.status(403).json({ message: 'unauthorized' });
+    }
     await DB.deleteFranchise(franchiseId);
     const authHeader = req.headers.authorization || '';
     res.locals.auth = authHeader.split(' ')[1] || null;
